+++ src/photosage/metadata/__init__.py
@@ -1,1 +1,15 @@
+from photosage.metadata.exif_reader import PhotoMetadata, extract_metadata, extract_photo_metadata
+from photosage.metadata.filename_from_metadata import date_from_metadata, filename_from_metadata
+from photosage.metadata.metadata_score import MetadataScore, metadata_is_sufficient, score_metadata, score_metadata_details
 
+__all__ = [
+    "MetadataScore",
+    "PhotoMetadata",
+    "date_from_metadata",
+    "extract_metadata",
+    "extract_photo_metadata",
+    "filename_from_metadata",
+    "metadata_is_sufficient",
+    "score_metadata",
+    "score_metadata_details",
+]